Split data
