+++ src/logic/chord/ChordGenerator.test.ts
