+++ src/logic/interval/data.ts
@@ -1,0 +1,15 @@
+// prettier-ignore
+export const INTERVAL_NUMBERS = new Map<string, number>([
+  ['1' , 0 ],
+  ['b2', 1 ], ['b9' , 1],
+  ['2' , 2 ], ['9'  , 2],
+  ['b3', 3 ], ['#9' , 3],
+  ['3' , 4 ],
+  ['4' , 5 ], ['11' , 5],
+  ['b5', 6 ], ['#11', 6],
+  ['5' , 7 ],
+  ['b6', 8 ], ['b13', 8], ['#5' , 8],
+  ['6' , 9 ], ['13' , 9], ['bb7', 9],
+  ['b7', 10],
+  ['7' , 11]
+]);
