+++ src/logic/note/data.ts
@@ -1,0 +1,1 @@
+export const NOTE_NAMES = ['C', 'C#', 'D', 'Eb', 'E', 'F', 'F#', 'G', 'G#', 'A', 'Bb', 'B'];
